Implementation of clean architecture and linter rules added
