+++ analysis_options.yaml
@@ -22,8 +22,17 @@ linter:
   # `// ignore_for_file: name_of_lint` syntax on the line or in the file
   # producing the lint.
   rules:
-    # avoid_print: false  # Uncomment to disable the `avoid_print` rule
-    # prefer_single_quotes: true  # Uncomment to enable the `prefer_single_quotes` rule
+    avoid_redundant_argument_values: true
+    prefer_const_constructors: true
+    prefer_const_declarations: true
+    prefer_relative_imports: true
+    depend_on_referenced_packages: false
+    library_private_types_in_public_api: false
+    use_build_context_synchronously: false
+    avoid_unnecessary_containers: true
+    void_checks: true
+    unnecessary_const: true
+    eol_at_end_of_file: true
 
 # Additional information about this file can be found at
 # https://dart.dev/guides/language/analysis-options
