--- lib/interfaces/i_repository.dart
@@ -1,3 +1,0 @@
-abstract class IMyRepository<T> {
-  Future<T> getData();
-}
