+++ lib/src/domain/model/movie.dart
