+++ lib/src/domain/repository/interfaces/i_genre_repository.dart
@@ -1,0 +1,8 @@
+abstract class IGenreRepository<T> {
+  Future<T> getData();
+
+  T getNameFromIds(
+    List<int> genresIds,
+    T genres,
+  );
+}
