+++ lib/src/domain/repository/interfaces/i_movie_repository.dart
@@ -1,0 +1,3 @@
+abstract class IMovieRepository<T> {
+  Future<T> getData();
+}
