+++ lib/src/presentation/widgets/popup_menu_item.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import '../../core/util/constants.dart';
+
+class AppBarPopUpItemRow extends StatelessWidget {
+  const AppBarPopUpItemRow({
+    super.key,
+    required this.name,
+    required this.icon,
+  });
+
+  final String name;
+  final IconData icon;
+  final TextStyle popupItemsTextStyle = const TextStyle(
+    fontFamily: ConstantsClass.tekturFontFamily,
+    fontWeight: FontWeight.bold,
+    fontSize: ConstantsClass.homeIconFontSize,
+    color: Colors.black,
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      children: [
+        Icon(
+          color: Colors.black,
+          icon,
+        ),
+        Text(
+          name,
+          style: popupItemsTextStyle,
+        ),
+      ],
+    );
+  }
+}
