--- lib/widgets/back_button.dart
@@ -1,21 +1,0 @@
-import 'package:flutter/material.dart';
-import '/shared/constants.dart';
-
-class MyBackButton extends StatelessWidget {
-  const MyBackButton({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return FloatingActionButton(
-      backgroundColor: ConstantsClass.appColor,
-      onPressed: () {
-        Navigator.pop(context);
-      },
-      heroTag: null,
-      child: const Icon(
-        color: Colors.white,
-        Icons.arrow_back,
-      ),
-    );
-  }
-}
